Put write strategies into own module
